requirements.txt
# GitHub API
PyGithub==2.1.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1

# Visualization
plotly==5.18.0
kaleido==0.2.1

# Template Engine
Jinja2==3.1.2

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
tqdm==4.66.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
